+++ tests/selection-traversal.rs
@@ -234,5 +234,4 @@ fn test_doc_uppercase() {
         all_matched,
         "<li>1</li><li>2</li><li>3</li><li>4</li><li>5</li><li>6</li>"
     );
-
 }