+++ tests/selection-manipulation.rs
@@ -9,7 +9,7 @@ fn test_replace_with_html() {
 
     println!("{}", doc.html());
 
-    let mut sel = doc.select("#main,#foot");
+    let sel = doc.select("#main,#foot");
     println!("{}", sel.length());
     sel.replace_with_html(r#"<div id="replace"></div>"#);
 
@@ -22,7 +22,7 @@ fn test_replace_with_html() {
 #[test]
 fn test_set_html() {
     let doc = doc_with_siblings();
-    let mut q = doc.select("#main, #foot");
+    let q = doc.select("#main, #foot");
     q.set_html(r#"<div id="replace">test</div>"#);
 
     assert_eq!(doc.select("#replace").length(), 2);
@@ -35,7 +35,7 @@ fn test_set_html() {
 #[test]
 fn test_set_html_no_match() {
     let doc = doc_with_siblings();
-    let mut q = doc.select("#notthere");
+    let q = doc.select("#notthere");
     q.set_html(r#"<div id="replace">test</div>"#);
     assert_eq!(doc.select("#replace").length(), 0);
 }
@@ -43,7 +43,7 @@ fn test_set_html_no_match() {
 #[test]
 fn test_set_html_empty() {
     let doc = doc_with_siblings();
-    let mut q = doc.select("#main");
+    let q = doc.select("#main");
     q.set_html("");
     assert_eq!(doc.select("#main").length(), 1);
     assert_eq!(doc.select("#main").children().length(), 0);
@@ -54,7 +54,7 @@ fn test_replace_with_selection() {
     let doc = doc_with_siblings();
 
     let s1 = doc.select("#nf5");
-    let mut sel = doc.select("#nf6");
+    let sel = doc.select("#nf6");
 
     sel.replace_with_selection(&s1);