+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to the `dom_query` crate will be documented in this file.
 ## [Unreleased]
 
 ## Added
-- Added `Selection::remove_attrs` method, a convenient way to remove multiple attributes from an element.
+- Added `Selection::remove_attrs` method, a convenient way to remove multiple attributes from all elements in the selection.
+- Added `Selection::rename` method, which allows to change all elements' names in the selection.
 
 ## Changed
 - Minor refactor of the codebase.