+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the `dom_query` crate will be documented in this file.
 ## Changed
 - Exposed `Document::tree`.
 - `Selection` methods that required `&mut` now doesn't require `&mut`.
+- Changed the project structure, now modules are divided based on the struct implementations.
 
 ## Added
 - Added `Node::append_html` and `Node::set_html` methods for creating children nodes of a single selected node.