+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 All notable changes to the `dom_query` crate will be documented in this file.
 
-## [Unreleased]
+## [0.6.0] - 2024-10-19
 
 ## Changed
 - Exposed `Document::tree`.
 - `Selection` methods that required `&mut` now doesn't require `&mut`.
-- Changed the project structure, now modules are divided based on the struct implementations.
+- Changed the project structure, now modules are divided based on the `struct` implementations.
 
 ## Added
 - Added `Node::append_html` and `Node::set_html` methods for creating children nodes of a single selected node.